refactor(cmd): split each command into its own file
Move the version, scan, and rules commands out of main.go into version.go,
scan.go, and rules.go (llm.go and mcp.go already followed this pattern). main.go
now holds only the root-command wiring, build metadata, and the exitCodeError
type. Pure code move — no behavior change; build, vet, gofmt, and the full test
suite are unchanged.
