+++ cmd/trustabl/rules.go
@@ -1,0 +1,53 @@
+package main
+
+import (
+	"errors"
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra"
+
+	"github.com/trustabl/trustabl/internal/rules"
+	"github.com/trustabl/trustabl/internal/rulesource"
+)
+
+func newRulesCommand() *cobra.Command {
+	rulesCmd := &cobra.Command{
+		Use:   "rules",
+		Short: "Manage Trustabl's detection rules",
+	}
+
+	var repo, ref string
+	pull := &cobra.Command{
+		Use:   "pull",
+		Short: "Download the detection rule packs into the local cache",
+		Args:  cobra.NoArgs,
+		RunE: func(_ *cobra.Command, _ []string) error {
+			if repo == "" {
+				repo = os.Getenv("TRUSTABL_RULES_REPO")
+			}
+			res, err := rulesource.Pull(
+				rulesource.Config{RepoURL: repo, Ref: ref},
+				rules.SupportedSchemaVersion,
+			)
+			if err != nil {
+				if errors.Is(err, rulesource.ErrNoCompatibleRules) {
+					return fmt.Errorf("the rules at the requested ref are newer than this "+
+						"Trustabl build can evaluate (this engine supports rule schema "+
+						"version up to %d); upgrade Trustabl or pull an older --rules-ref",
+						rules.SupportedSchemaVersion)
+				}
+				return fmt.Errorf("rules pull: %w", err)
+			}
+			fmt.Printf("Pulled rules from %s at %s\n", res.RepoURL, res.SHA)
+			return nil
+		},
+	}
+	pull.Flags().StringVar(&repo, "rules-repo", "",
+		"rules repository URL (default: official trustabl-rules; or TRUSTABL_RULES_REPO)")
+	pull.Flags().StringVar(&ref, "rules-ref", "",
+		"rules branch or tag to pull (default: the repo's default branch)")
+
+	rulesCmd.AddCommand(pull)
+	return rulesCmd
+}
